Extract ApplicationIdentity validations
We had some validations in application_identity.rb and some in
validate_application_identity.rb. This was confusing to the reader so this
commit fixes that and makes the code clearer.
